25: Yellow flash Battle transition
First, use `CanvasLayer` to ensure it follows the camera. Set a `ColorRect` as child.
Second, ensure its `layer` is 100 to make it appear on top.
Third, in `AnimationPlayer`, set keys for `ColorRect` transition from white to yellow.
Fourth, in `AnimationPlayer`, set modulate from 255 to 0.
Fifth, give `ColorRect` a new canvas material and set it to `add`.
Finally, key the `CanvasLayer` visibility to visible and turn it off for ease of look.