52: Load `elizabeth_stats` into `ReferenceStash`
This way, `elizabeth_stats` will still be shared among all Godot nodes when
they decide to access it, but it is never freed when `Battle` scene exits
(where `PlayerBattleUnit`, which needs `elizabeth_stats`, also exits).
This ensures that the `elizabeth_stats` is always loaded during the game
and thus stores the level progression during the game.